tools/python/requirements.txt
PyYAML==6.0.2
requests==2.32.3
pyjson==1.4.1
ics==0.7.2
matplotlib